Move transaction stuff into the events module
